+++ common/core/desktop/src/kmx/kmx_consts.cpp
@@ -1,5 +1,5 @@
 
-#include "kmx_processor.h"
+#include "kmx_processevent.h"
 
 namespace km {
 namespace kbp {
