+++ common/core/desktop/src/kmx/kmx_xstring.cpp
@@ -6,7 +6,7 @@
 #include <iterator>
 #include <codecvt>
 #include <locale>
-#include "kmx_processor.h"
+#include "kmx_processevent.h"
 #include "utfcodec.hpp"
