+++ common/core/desktop/tests/unit/kmx/kmx.cpp
@@ -17,7 +17,7 @@
 #include <string>
 #include <type_traits>
 
-#include <kmx/kmx_processor.h>
+#include <kmx/kmx_processevent.h>
 
 #include "path.hpp"
 #include "state.hpp"
